Switch to a linked hash map for storing running totals of daily balances
Using a LinkedHashMap rather than a TreeMap means that the act of calculating all daily balances is now an O(n) operation rather than O(n * log n). While the preceding list sort already made this method O(n * log n) in either case, it is a minor change to improve the efficiency of the actual calculation operation.

The more significant change may be how the missing days are filled in; previously it relied upon the `lastOption` method to determine what was the last day, but this is actually an O(log n) operation in the case of a TreeMap and, even worse, an O(n) operation in the case of a LinkedHashMap. Instead of incurring the cost of calling `lastOption` on every single iteration (potentially making the efficiency of calculating all balances O(n * log n * log n) or O(n * n)), the previous day's entry is simply carried forward to each iteration of the fold operation, making its efficiency bound to the number of missing days rather than the number of missing days plus the number of preceding balance entries.