Switch from a stream of transaction pages to a list
The use of a Stream when collecting the transaction pages was originally intended to provide a means for a caller to be able to operate on a single page at a time without having to hold the entire collection of pages in memory at once. However, that solution was not feasible because of the fact that the transactions must be sorted into chronological order before they can be tallied, so it was abandoned mid-implementation. The type remained Stream even though the solution was unable to take advantage of it. Switch to a Seq to eliminate confusion caused by the presence of the Stream return type.